refactor #138 media, recommend 바인딩 분리
