+++ lib/features/main/main_binding.dart
@@ -1,4 +1,5 @@
-import 'package:debateseason_frontend_v1/features/home/bindings/home_binding.dart';
+import 'package:debateseason_frontend_v1/features/home/bindings/media_binding.dart';
+import 'package:debateseason_frontend_v1/features/home/bindings/recommend_binding.dart';
 import 'package:debateseason_frontend_v1/features/issuemap/bindings/issuemap_binding.dart';
 import 'package:debateseason_frontend_v1/features/profile/bindings/profile_binding.dart';
 import 'package:get/get.dart';
@@ -9,7 +10,8 @@ class MainBinding extends Bindings {
   @override
   void dependencies() {
     Get.lazyPut(() => MainViewModel());
-    HomeBinding().dependencies();
+    MediaBinding().dependencies();
+    RecommendBinding().dependencies();
     IssuemapBinding().dependencies();
     ProfileBinding().dependencies();
   }
